requirements.txt
numpy==1.15.3
scipy==0.19.0
matplotlib==2.0.0
scikit-image
opencv-python
torch torchvision
Ipython
tensorboard-pytorch --no-cache-dir
tensorflow-gpu==1.5
